+++ src/main/java/io/secuconnect/client/model/LoyaltyMerchantcardsProductWithReceiptModel.java
@@ -23,6 +23,12 @@ public class LoyaltyMerchantcardsProductWithReceiptModel extends LoyaltyMerchant
   @SerializedName("receipt")
   private List<ReceiptTypeValue> receipt = null;
 
+  @SerializedName("trans_id")
+  private Integer transId = null;
+
+  @SerializedName("amount")
+  private Integer amount = null;
+
   public LoyaltyMerchantcardsProductWithReceiptModel receiptHeader(List<ReceiptType> receiptHeader) {
     this.receiptHeader = receiptHeader;
     return this;
@@ -73,6 +79,40 @@ public class LoyaltyMerchantcardsProductWithReceiptModel extends LoyaltyMerchant
     this.receipt = receipt;
   }
 
+  public LoyaltyMerchantcardsProductWithReceiptModel transId(Integer transId) {
+    this.transId = transId;
+    return this;
+  }
+
+   /**
+   * Created TransactionId
+   * @return transId
+  **/
+  public Integer getTransId() {
+    return transId;
+  }
+
+  public void setTransId(Integer transId) {
+    this.transId = transId;
+  }
+
+  public LoyaltyMerchantcardsProductWithReceiptModel amount(Integer amount) {
+    this.amount = amount;
+    return this;
+  }
+
+   /**
+   * Amount
+   * @return amount
+  **/
+  public Integer getAmount() {
+    return amount;
+  }
+
+  public void setAmount(Integer amount) {
+    this.amount = amount;
+  }
+
   @Override
   public boolean equals(Object o) {
     if (this == o) {
@@ -84,12 +124,14 @@ public class LoyaltyMerchantcardsProductWithReceiptModel extends LoyaltyMerchant
     LoyaltyMerchantcardsProductWithReceiptModel loyaltyMerchantcardsProductWithReceiptModel = (LoyaltyMerchantcardsProductWithReceiptModel) o;
     return Objects.equals(this.receiptHeader, loyaltyMerchantcardsProductWithReceiptModel.receiptHeader) &&
         Objects.equals(this.receipt, loyaltyMerchantcardsProductWithReceiptModel.receipt) &&
+        Objects.equals(this.transId, loyaltyMerchantcardsProductWithReceiptModel.transId) &&
+        Objects.equals(this.amount, loyaltyMerchantcardsProductWithReceiptModel.amount) &&
         super.equals(o);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(receiptHeader, receipt, super.hashCode());
+    return Objects.hash(receiptHeader, receipt, transId, amount, super.hashCode());
   }
 
   @Override
@@ -99,6 +141,8 @@ public class LoyaltyMerchantcardsProductWithReceiptModel extends LoyaltyMerchant
     sb.append("    ").append(toIndentedString(super.toString())).append("\n");
     sb.append("    receiptHeader: ").append(toIndentedString(receiptHeader)).append("\n");
     sb.append("    receipt: ").append(toIndentedString(receipt)).append("\n");
+    sb.append("    transId: ").append(toIndentedString(transId)).append("\n");
+    sb.append("    amount: ").append(toIndentedString(amount)).append("\n");
     sb.append("}");
     return sb.toString();
   }